+++ poulpy-cpu-avx512/src/ntt4x30_avx512/module.rs
@@ -44,7 +44,7 @@ pub struct NTT4x30Avx512Handle {
 impl poulpy_hal::execution::ScratchWorkers for NTT4x30Avx512 {}
 
 impl Backend for NTT4x30Avx512 {
-    const DFT_MAX_BITS: usize = <poulpy_cpu_ref::NTT4x30Ref as Backend>::DFT_MAX_BITS;
+    const DFT_MAX_BITS: f64 = <poulpy_cpu_ref::NTT4x30Ref as Backend>::DFT_MAX_BITS;
     const DFT_LIMBS_CONTIGUOUS: bool = true;
 
     type TaskExecutor = poulpy_hal::execution::SerialTaskExecutor;
