+++ poulpy-cpu-ref/src/ntt4x30/module.rs
@@ -46,7 +46,7 @@ pub struct NTT4x30RefHandle {
 impl poulpy_hal::execution::ScratchWorkers for NTT4x30Ref {}
 
 impl Backend for NTT4x30Ref {
-    const DFT_MAX_BITS: usize = Primes30::LOG_Q_PRODUCT;
+    const DFT_MAX_BITS: f64 = Primes30::LOG_Q_PRODUCT;
     const DFT_LIMBS_CONTIGUOUS: bool = true;
 
     type TaskExecutor = poulpy_hal::execution::SerialTaskExecutor;
